Добавление модульности
