+++ js/data.js
@@ -1,0 +1,45 @@
+const allMessages = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const allNames = [
+  'Артём',
+  'Алексей',
+  'Мария',
+  'Светлана'
+];
+
+const descriptions = [
+  'Бассейн',
+  'Указатель',
+  'Лазурная вода',
+  'Не твоя',
+  'Еда',
+  'Авто',
+  'Клубника',
+  'Компот',
+  'Самолёт',
+  'Бархатные тяги',
+  'Дорога к морю',
+  'Авто поменьше',
+  'Салат',
+  'Живодёрство',
+  'Подкрадули',
+  'Вид из иллюминатора',
+  'Хор',
+  'Кучерявое ретро',
+  'Подглядули',
+  'Пальмы',
+  'Еда посытнее',
+  'Закат',
+  'Членистоногое',
+  'Концерт',
+  'Что-то происходит'
+];
+
+export {allMessages, allNames, descriptions};
